+++ twitchService/src/main/java/cersei/twitchservice/model/Viewer.java
@@ -22,5 +22,5 @@ public class Viewer {
 
     LocalDateTime dateTime;
 
-    long viewers;
+    int viewers;
 }
